requirements.txt
discord.py
jieba
python-dotenv
transformers
torch